pyproject.toml
[project]
name = "pix2pix-gan-from-scratch-pytorch"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "numpy>=2.0.2",
    "pillow>=11.3.0",
    "torch>=2.8.0",
]
